I changed Excel handling
I added convert_excel_to_csv_2 and _3 as well as _primitive_read_excel_2 and _3.  Logic: convert_exce_to_csv_1 and _primitive_read_excel_1 are for initially displaying the file before user is asked about the header.  Then _2 is used if the file does have a header - first row is tested for empty cells.  _3 is used if the header needs to be appended.